+++ core/src/dird/job.cc
@@ -173,7 +173,6 @@ bool SetupJob(JobControlRecord* jcr, bool suppress_output)
     goto bail_out;
   }
   Dmsg0(150, "DB opened\n");
-  if (!jcr->dir_impl->fname) { jcr->dir_impl->fname = GetPoolMemory(PM_FNAME); }
 
   if (!jcr->dir_impl->res.pool_source) {
     jcr->dir_impl->res.pool_source = GetPoolMemory(PM_MESSAGE);
@@ -1495,7 +1494,6 @@ void DirdFreeJcrPointers(JobControlRecord* jcr)
   FreeAndNullPoolMemory(jcr->JobIds);
   FreeAndNullPoolMemory(jcr->dir_impl->client_uname);
   FreeAndNullPoolMemory(jcr->attr);
-  FreeAndNullPoolMemory(jcr->dir_impl->fname);
 }
 
 /**
@@ -1540,7 +1538,6 @@ void DirdFreeJcr(JobControlRecord* jcr)
   }
 
   FreeAndNullPoolMemory(jcr->starttime_string);
-  FreeAndNullPoolMemory(jcr->dir_impl->fname);
   FreeAndNullPoolMemory(jcr->dir_impl->res.pool_source);
   FreeAndNullPoolMemory(jcr->dir_impl->res.npool_source);
   FreeAndNullPoolMemory(jcr->dir_impl->res.rpool_source);
@@ -1550,6 +1547,7 @@ void DirdFreeJcr(JobControlRecord* jcr)
   FreeAndNullPoolMemory(jcr->dir_impl->FDSecureEraseCmd);
   FreeAndNullPoolMemory(jcr->dir_impl->SDSecureEraseCmd);
   FreeAndNullPoolMemory(jcr->dir_impl->vf_jobids);
+  if (jcr->dir_impl->plugin_options) { free(jcr->dir_impl->plugin_options); }
 
   // Delete lists setup to hold storage pointers
   FreeRwstorage(jcr);
@@ -1628,7 +1626,6 @@ void SetJcrDefaults(JobControlRecord* jcr, JobResource* job)
       break;
   }
 
-  if (!jcr->dir_impl->fname) { jcr->dir_impl->fname = GetPoolMemory(PM_FNAME); }
   if (!jcr->dir_impl->res.pool_source) {
     jcr->dir_impl->res.pool_source = GetPoolMemory(PM_MESSAGE);
     PmStrcpy(jcr->dir_impl->res.pool_source, T_("unknown source"));