+++ core/src/lib/parse_conf_state_machine.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2024 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -71,6 +71,11 @@ bool ConfigParserStateMachine::ParseAllTokens()
           case ParseInternalReturnCode::kGetNextToken:
             continue;
           case ParseInternalReturnCode::kError:
+            // delete the inited resource
+            my_config_.FreeResourceCb_(
+                currently_parsed_resource_.allocated_resource_,
+                currently_parsed_resource_.rcode_);
+            currently_parsed_resource_.allocated_resource_ = nullptr;
             return false;
           default:
             ASSERT(false);